requirements.txt
numpy==2.1.1
numpy-stl==3.1.2
scipy==1.14.1
